requirements.txt
anyio==3.6.2
certifi==2023.5.7
cffi==1.15.1
click==8.1.3
cryptography==40.0.2
dnspython==2.3.0
email-validator==2.0.0.post2
fastapi==0.95.1
h11==0.14.0
httpcore==0.17.0
httptools==0.5.0
httpx==0.24.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
orjson==3.8.12
pycparser==2.21
pydantic==1.10.7
PyJWT==2.7.0
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0
secure==0.3.0
sniffio==1.3.0
SQLAlchemy==1.4.41
sqlalchemy2-stubs==0.0.2a34
sqlmodel==0.0.8
starlette==0.26.1
typing_extensions==4.5.0
ujson==5.7.0
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
